move WellFromedInstrument to relax.ir.instrument
